requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
openai==1.6.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
drf-nested-routers==0.93.4
pydub==0.25.1
setuptools>=68.0.0 